requirements.txt
boto3==1.26.135
simple-salesforce==1.12.4
requests==2.30.0
aws-lambda-powertools==2.16.2
python-dotenv==1.0.0

pyarrow==12.0.0
fastparquet==2023.4.0
numpy==1.24.3
pandas==1.5.3